+++ src/utils/markfluence/elements/break.ts
@@ -1,0 +1,4 @@
+import type { Break } from "mdast";
+import { registerConverter } from "../registry.js";
+
+registerConverter<Break>("break", () => "<br/>");