+++ src/utils/markfluence/elements/html.ts
@@ -1,0 +1,5 @@
+import type { Html } from "mdast";
+import { registerConverter } from "../registry.js";
+import { postProcessStorageHtml } from "../utils.js";
+
+registerConverter<Html>("html", (node) => postProcessStorageHtml(node.value));