+++ src/utils/markfluence/elements/root.ts
@@ -1,0 +1,4 @@
+import type { Root } from "mdast";
+import { registerConverter } from "../registry.js";
+
+registerConverter<Root>("root", (node, _context, convertChildren) => convertChildren(node));